feat(ci): add nightly build script
Add CI to create a nightly release of Conjure Oxide.

This performs a build check on Conjure Oxide for all platforms, then on
success, uploads zip files containing prebuilt binaries of Conjure Oxide
and its dependencies to a GitHub release.

MOTIVATION

The primary motivation for this is to provide up to date builds for
long-running, nightly CI checks. In particular, I wish to track Conjure
Oxide's compatibility with a large folder of Essence Prime models.

Having the latest version of Conjure Oxide and its dependencies in a
bundled together downloadable from a known URL makes installing it in
such scripts easier: just `wget` the release and `unzip` it.

A secondary advantage is providing pre-compiled binaries for beta
testing.

DETAILS

This CI generates the following releases for each platform:

  + standalone - just conjure_oxide

  + with-conjure - conjure_oxide and conjure

  + with-solvers - conjure_oxide, and the contents of Conjures'
      with-solvers release.

The following platforms are used:

  - x86_64-darwin
  - aarch64-darwin
  - x86_64-linux
  - aarch64-linux

Each nightly build overrides the `nightly` release on Github,
overwriting the release files in it, and updating the description with
the current date.

Unlike the main CI, this release checks builds for more platforms and
does a clean build (instead of using sccache).

In the future, this workflow could contain CI jobs that should be done
nightly using these builds - e.g. performance testing, longer test
suites.