feat(ast): add `range!` and `domain_int!` macros
Add `range!` and `domain_int!` macros to simplify the creation of
integer domains in the code.

For example, we can now write the following:

```
domain_int!(1,2,3..a**2,10..)
```

where a is a rust variable.

My motivation for this was to reduce boilerplate in `HasDomain`
implementations.

I realise after writing this that this somewhat overlaps with Georgii's
Essence macro work. However, that currently doesn't support domains. As
domains don't currently take Essence expressions as input, we don't
really need the full power of Georgii's procedural macro and
quasiquotation, etc, having Rust expressions be substitutable into the
macro (as they are here) is enough.