feat!: make references store a shared pointer to their declarations
Change the type of references from `Atom::Reference(Name)` to
`Atom::Reference(Name,Rc<RefCell<Declaration>>)`.

To make declaration pointers both shared and mutable, this commit also
changes the type of declaration pointers from `Rc<Declaration>` to
`Rc<RefCell<Declaration>>`.

This has consequences for the parser (PARSER CHANGES) and
(de)serialisation (SERIALISATION / DESERIALISATION). In particular, we
now require all referenced names to be in the symbol table at
parse-time.

This commit only contains the type signature change, and any fixes
needed to make the tests pass. More work is needed to clean this up and
make everything fully bug free. See issue #910 for details.

Importantly, this commit breaks some tests that use `essence_expr!`. To
keep this commit from getting any larger, we've turned off these tests
for now and will fix them in future commits.

The interesting parts of this patch are in `conjure_core` - changes made
in `conjure_rules` and other crates are mainly mechanical refactors of
usages of `Atom::Reference`.

PARSER CHANGES

We parse all declarations first. Then, when parsing a reference, we
lookup the declaration pointer associated with that name in the current
scope, and use that in the constructed reference.

This means that we had to make sure everything we reference is in the
symbol table at parse time:

* References can contain record field names, so this patch adds
  `DeclarationKind::RecordField` to store them in. For example:

  ```
  find x: record {A: int(0..1), B: int(0..2)}
  such that

  x[A] = 0
  ```

* When building a comprehension return expression, we need access to the
  declarations of the generators inside the comprehension. Before this
  commit, we only built a comprehensions symbol table after the return
  expression was parsed. This commit adds
  `ComprehensionBuilder::symbol_table`, which returns a symbol table
  containing the declarations in the comprehension so far.

SERIALIZATION / DESERIALIZATION

As they now contain `Rc RefCell`'s, we now serialize and deserialize
references using ids in a similar manner to symbol tables.

We serialize a reference to some variable `x` by replacing its
declaration pointer with the unique id of that declaration:

```
"Reference" : [
  {"User": "x"},
  316
]
```

When we serialize the symbol table entry for the declaration, we store
the full declaration:

```
{"name": {"User": "x"}, "id": 316, "kind": { "DecisionVariable: { ... } }}
```

To deserialize the preceding reference, we first create a dummy
declaration for it with the correct id:

```
Atom::Reference(User("x"), Rc::new(RefCell::new(Declaration::default_with_id(316))))
```

Then, we gather the serialised symbol tables, find the declaration with
id 316, and replace the dummy declaration with a pointer to the real
declaration.

This process is done in `SerdeModel::initialise()` in
`crates/conjure_core/src/ast/model.rs`.

For more details on how we serialize and deserialize Rc Refcells, and
why we do it this way, see 863643260e (feat!: add parent symbol tables
(#680), 2025-02-17).

Tracking-issue: #910
Co-authored-by: Niklas Dewally <niklas@dewally.com>
Commit-message-by: Niklas Dewally <niklas@dewally.com>