feat(ast): make `DeclarationPtr::domain` return `Option<Domain>`
Make `DeclarationPtr::domain` return `Option<Domain>` instead of
`Option<Ref<Domain>>`.

This is necessary for a future commit in which I define the domain of a
value letting. In that commit, I tried to call `Expression::domain_of`
inside `DeclarationPtr::domain`. To fit in the return type of
`DeclarationPtr::domain`, it had to return a borrow of this domain;
however, that caused a compiler error due to returning a temporarily
borrowed value.

This also make `Declaration::domain`'s return type more similar to the
other domain functions.