feat: better comprehension expansion
Add better comprehension expansion for ac operators ("expand-ac"). This
performs pruning on a comprehension by passing the generator model to
Minion.

This is enabled by the `Better_Comprehension_Expansion` ruleset. This
ruleset is on by default,  but can be turned off with the
`--no-use-expand-ac` CLI flag.

For more details, please read our (upcoming) modref paper "Solver-aided
expansion of loops"