feat!: move `conjure_oxide` to `crates/conjure-cp-cli`
Rename `conjure_oxide` crate to `conjure-cp-cli`, and move it into the
`crates` folder.

This name makes it clearer that this crate just defines
the CLI and nothing else.

* Move integration tests to their own crate, `tests-integration`. It
  does not make sense to keep the integration tests with the CLI.
  Furthermore, this reduces rebuilds of `conjure_oxide`, as the
  integration tests' `build.rs` script requires a full crate rebuild
  every time a test file changes (including generated output!).

* Rename conjure oxide executable from `conjure_oxide` to
  `conjure-oxide`.

MOTIVATION

Before `conjure-cp` was added, `conjure_oxide` used to act as a library
facade, as well as providing the CLI. This is no longer the case. This
commit refactors `conjure_oxide` to just contain CLI related code,
moving all library related stuff to `conjure-cp`.

FUTURE WORK

In this commit, I have not changed any of the utility modules inside
`conjure-cp-cli`. It is possible that some of these are only used by the
integration tester, so should be moved to that crate instead.