chore(.gitignore): ignore directories or symlinks with the same name (#1076)
Before this PR, for example `target/` would ignore a directory called
target but it wouldn't ignore a symlink called `target`. If the trailing
`/` is omitted git ignores both the directory or a symlink with the same
name.

--

Just adds target file (symlink) to the .gitignore

This links to the post I made on Teams:

```
Hey! A couple of people were running into the problem of quota exceeded with conjure-oxide. One of the reasons could be that the target in the root of conjure-oxide folder is getting into snapshot. Shikhar Srivastava solution to that was to create a symlink to some folder that is not getting saved (e.g Downloads). 
 
That's what I've done
 
~Documents/VIP/conjure-oxide $ ln -sf /cs/home/yb33/Downloads/ ./target
```