Tightening Rules on Included Macros (#1102)
## Description
This PR aims to better define which macros we wish to include in our
coverage testing.
The given example is the Derive macro, which creates new functions that
do not need tested, yet get counted in the overall coverage score.

## Related Issues
This PR is related to issue #1101 and discussion #1086

## Key Changes
- Added the derive macro to GRCOV_EXCLUDE_LINES in `tools/coverage.sh`

## How to Test
Run `tools/coverage.sh` and note what is included in GRCOV_EXCLUDE_LINES