Add a Coverage rule for the Makefile (#1054)
Based the desired implementation from #827 and #1052 in which we use the
coverage results to direct the implementation of new tests and
additional code, I created a simple rule to make sure of the following:
1. That people that want to check for the test coverage have access to
the necessary dependencies (GRCov and LLVM_TOOLS).
2. That they can produce detailed coverage reports repeatedly in a clean
and easy manner that abstracts over the actual commands.